requirements.txt
torch
torchvision
ultralytics
opencv-python
pyyaml
tqdm